chore: convert repo to a mono-package #4
