+++ .gitignore
@@ -1,30 +1,14 @@
-# Compiled JS lib code
-**/lib
-
-# Dependency directories
-node_modules/
-**/node_modules
-
-# Logs
+dist
+node_modules
 logs
 *.log
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
-
-# TypeScript cache
 *.tsbuildinfo
-
-# Optional npm cache directory
 .npm
-
-# Optional eslint cache
 .eslintcache
-
-# dotenv environment variables file
 .env
 .env.test
-
-# Weird mac file.
 .DS_Store
