+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  automock: false,
+  setupFiles: ['./test/setupJest.ts'],
+};
