+++ package.json
@@ -1,19 +1,48 @@
 {
-  "name": "root",
+  "name": "@airgrid/edgekit",
+  "version": "0.0.0-dev.0",
+  "homepage": "https://edgekit.org/",
+  "author": "AirGrid <https://airgrid.io>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/AirGrid/edgekit.git"
+  },
+  "description": "A privacy focused library for cookie-less audience creation.",
+  "keywords": [
+    "privacy",
+    "cookie-less",
+    "audiences",
+    "programmatic",
+    "edgekit",
+    "prebid"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
-    "build": "tsc -b packages",
-    "watch": "tsc -w packages",
-    "lint": "eslint packages/**/src packages/**/__test__ --ext js,ts",
-    "lint:fix": "eslint packages/**/src packages/**/__test__ --ext js,ts --fix"
+    "clean": "rimraf dist",
+    "compile": "tsc --traceResolution",
+    "build": "run-s clean compile",
+    "watch": "tsc -w",
+    "test": "jest --verbose",
+    "lint": "eslint src test --ext js,ts",
+    "lint:fix": "eslint src test --ext js,ts --fix"
   },
   "devDependencies": {
+    "@types/jest": "^26.0.3",
     "@typescript-eslint/eslint-plugin": "^3.3.0",
     "@typescript-eslint/parser": "^3.3.0",
     "eslint": "^7.3.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "lerna": "^3.22.1",
+    "jest": "^26.1.0",
+    "jest-fetch-mock": "^3.0.3",
+    "npm-run-all": "^4.1.5",
     "prettier": "^2.0.5",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.1.1",
+    "tslib": "^2.0.0",
     "typescript": "^3.9.5"
   }
 }
