--- packages/audiences/src/definitions/interest/travel.ts
@@ -1,29 +1,0 @@
-import { ICondition } from '@edgekit/types';
-
-const condition: ICondition = {
-  filter: {
-    queries: [
-      {
-        property: 'keywords',
-        value: ['beach', 'holiday'],
-      },
-    ],
-  },
-  rules: [
-    {
-      reducer: {
-        name: 'count',
-      },
-      matcher: {
-        name: 'gt',
-        args: 2,
-      },
-    },
-  ],
-};
-
-export default {
-  id: 'travel',
-  ttl: 3600,
-  conditions: [condition],
-};
