--- packages/audiences/src/index.ts
@@ -1,4 +1,0 @@
-import interestTravel from './definitions/interest/travel';
-import interestSport from './definitions/interest/sport';
-
-export default [interestTravel, interestSport];
