--- packages/core/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  automock: false,
-  setupFiles: [
-    './__test__/setupJest.ts',
-  ]
-};
