--- packages/engine/__test__/index.html
@@ -1,5 +1,0 @@
-<script src="http://127.0.0.1:8080/edkt.umd.development.js"></script>
-<script type="module">
-  console.log(edkt);
-  edkt.check();
-</script>
