--- packages/engine/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@edgekit/engine",
-  "version": "0.0.1-alpha.3",
-  "description": "",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "files": [
-    "/lib"
-  ],
-  "scripts": {
-    "compile": "tsc",
-    "test": "jest --verbose"
-  },
-  "author": "ydennisy",
-  "license": "MIT",
-  "dependencies": {
-    "@edgekit/types": "^0.0.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.0",
-    "jest": "^26.0.1",
-    "jest-fetch-mock": "^3.0.3",
-    "ts-jest": "^26.1.0",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.5"
-  },
-  "gitHead": "d7e3c229de3d6564757ce95af38ac27963af688e"
-}
