--- packages/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "files": [],
-  "references": [
-    { "path": "core" },
-    { "path": "engine" },
-    { "path": "types" },
-    { "path": "audiences" }
-  ]
-}
