--- packages/types/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@edgekit/types",
-  "version": "0.0.1-alpha.3",
-  "description": "",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "files": [
-    "/lib"
-  ],
-  "scripts": {
-    "compile": "tsc"
-  },
-  "author": "ydennisy",
-  "license": "MIT",
-  "devDependencies": {
-    "typescript": "^3.9.5"
-  },
-  "gitHead": "d7e3c229de3d6564757ce95af38ac27963af688e"
-}
