+++ src/audiences/index.ts
@@ -1,0 +1,4 @@
+import interestTravel from './interest/travel';
+import interestSport from './interest/sport';
+
+export default [interestTravel, interestSport];
