+++ src/audiences/interest/sport.ts
@@ -1,0 +1,29 @@
+import { ICondition } from 'types';
+
+const condition: ICondition = {
+  filter: {
+    queries: [
+      {
+        property: 'keywords',
+        value: ['sport', 'football'],
+      },
+    ],
+  },
+  rules: [
+    {
+      reducer: {
+        name: 'count',
+      },
+      matcher: {
+        name: 'gt',
+        args: 1,
+      },
+    },
+  ],
+};
+
+export default {
+  id: 'sport',
+  ttl: 3600,
+  conditions: [condition],
+};
