+++ src/utils/index.ts
@@ -1,0 +1,26 @@
+export const timeStampInSecs = (): number => Math.round(Date.now() / 1000);
+
+const get = (key: string): any => {
+  const value = localStorage.getItem(key);
+  if (!value) return undefined;
+
+  try {
+    return JSON.parse(value);
+  } catch (e) {
+    return undefined;
+  }
+};
+
+const set = (key: string, value: any): void => {
+  try {
+    const serialized = JSON.stringify(value);
+    localStorage.setItem(key, serialized);
+  } catch (e) {
+    // ignore...
+  }
+};
+
+export const storage = {
+  set,
+  get,
+};
