+++ tsconfig.json
@@ -1,0 +1,27 @@
+{
+  "include": ["src", "types", "test"],
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "dist",
+    "lib": ["dom", "esnext"],
+    "importHelpers": true,
+    "moduleResolution": "node",
+    "baseUrl": ".",
+    "paths": {
+      // "*": ["src/*", "types/*", "node_modules/*"],
+      "types/*": ["types/*"]
+    },
+    "composite": true,
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "esModuleInterop": true,
+    "target": "es5",
+    "module": "commonjs"
+  }
+}
